magit-wip-commit-worktree: Use --ignore-skip-worktree-entries
With the previous commit, the read-tree call no longer fails when a
skip-worktree entry in the main index differs from an entry in the wip
ref.  However, the treatment of a skip-worktree file in the wip ref
still has an issue: if such a file is passed explicitly to
magit-wip-commit-worktree (e.g., after saving the file), the file is
deleted in the wip ref, despite the fact that the skip-worktree bit
was copied over from the main index in the read-tree call.  At this
point, the skip-worktree bit is lost, and, the next time the file is
passed to magit-wip-commit-worktree, the update-index call will add
the file back.

This happens because `update-index --remove' by default doesn't
respect the skip-worktree bit.  Luckily, Git v2.25.0 added a new
option, --ignore-skip-worktree-entries, that disables this behavior.
Teach magit-wip-commit-worktree to pass that flag to update-index if a
recent enough Git version is detected.

Note that we could avoid this problem on older Git versions by
switching back to `git add', but then we would reintroduce the
problems resolved by 57f2d0f8 (magit-wip-commit-worktree: Avoid
failing git call on file deletion, 2020-02-06).  That doesn't seem
like a good trade-off because this skip-worktree issue (1) isn't
likely to be common, (2) doesn't lose information, and (3) will
resolve itself in time.

[0]: The reason for this is unclear, but it's not going to change
     given that update-index is plumbing.

     https://lore.kernel.org/git/xmqq36fda3i8.fsf@gitster-ct.c.googlers.com/